requirements.txt
# Core Dependencies
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.25
openai==1.45.0
httpx==0.27.2

# Vector Store
pymongo==4.8.0

# PDF Processing
pypdf==4.3.1
pdfplumber==0.11.4

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# Eval & Testing
numpy==1.26.4
pandas==2.2.2

# Optional: Progress bars
tqdm==4.66.5

# Embeddings - Local (sentence-transformers)
sentence-transformers==3.0.1

# Tiktoken for token counting
tiktoken==0.7.0
